Modify MandateTypeVC code
